+++ planx_geostats_provider.py
@@ -24,6 +24,7 @@ from .algorithms.alg_gwr import GWRAlgorithm
 from .algorithms.alg_mgwr import MGWRAlgorithm
 from .algorithms.alg_exploratory_regression import ExploratoryRegressionAlgorithm
 from .algorithms.alg_generalized_linear_regression import GeneralizedLinearRegressionAlgorithm
+from .algorithms.alg_model_comparison import ModelComparisonAlgorithm
 from .algorithms.alg_median_center import MedianCenterAlgorithm
 from .algorithms.alg_central_feature import CentralFeatureAlgorithm
 from .algorithms.alg_general_g import GeneralGAlgorithm
@@ -94,3 +95,4 @@ class PlanXGeoStatsProvider(QgsProcessingProvider):
         self.addAlgorithm(GWRAlgorithm())
         self.addAlgorithm(MGWRAlgorithm())
         self.addAlgorithm(SensitivityTestAlgorithm())
+        self.addAlgorithm(ModelComparisonAlgorithm())